// src/apis/models/MatchUserCreateRequestRole.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * MatchDay Swagger
 * MatchDay 요구사항에 대한 REST API
 * OpenAPI spec version: 0.0.1
 */

/**
 * 매치에서의 역할
 */
export type MatchUserCreateRequestRole =
  (typeof MatchUserCreateRequestRole)[keyof typeof MatchUserCreateRequestRole];

export const MatchUserCreateRequestRole = {
  ADMIN: 'ADMIN',
  START_PLAYER: 'START_PLAYER',
  SUB_PLAYER: 'SUB_PLAYER',
  ARCHIVES: 'ARCHIVES',
} as const;